fix: reduce clippy warnings, improve sync and WS shutdown handling
- Auto-fix 97 clippy issues (redundant closures, div_ceil, conversions)
- Replace eprintln!/println! with tracing macros (except CLI wallet output)
- Add tracing dependency to call-storage crate
- Block announcements now trigger sync request when peer is ahead
- Iterative batch sync: keeps requesting blocks until caught up (was single
  batch with 5 attempts); light client built once at sync start
- WS ServerHandle stored in CallNode for graceful shutdown (was detached)
- Fix light client test imports: BlockSignature, IncrementalMerkleTree,
  keccak256
- Fix &PathBuf -> &Path in function signatures
- Remove unnecessary identity map calls
- Handle unused Result values in agent executor

+++ crates/protocol/src/sponsor.rs
@@ -169,7 +169,7 @@ impl SponsorRegistry {
         }
 
         // Check pool balance
-        if (pool.balance as u128) < fee {
+        if pool.balance < fee {
             return Err(ProtocolError::SponsorError(
                 "insufficient pool balance".into(),
             ));
